separated pattern samples to files and assets
